fix(test): extract selectMobileAuthVariant to standalone module (CI env isolation)
PR #100 CI test fail の原因と修正:

## 原因
MobileAuthSection.test.ts が `selectMobileAuthVariant` を MobileAuthSection.tsx
から import していたが、MobileAuthSection.tsx は `useStore` 経由で store/index →
firebaseClient.ts まで transitive import する。CI には VITE_FIREBASE_* env が
無いため firebaseClient の必須 env チェック (line 18) で throw、test の import
段階で失敗 (0 test executed)。

ローカルでは .env.local の VITE_FIREBASE_* が読まれて PASS していたため未検出。
env -u VITE_FIREBASE_* で再現確認済み。

## 修正
selectMobileAuthVariant を `components/mobileAuthVariant.ts` 新規ファイルに分離:
- 型 (AuthStatus, CurrentUser) は store/authSlice からの type-only import なので
  値を pull せず firebaseClient 経路は通らない
- MobileAuthSection.tsx は新ファイルから import
- MobileAuthSection.test.ts も新ファイルから import (drift 防止維持)

## 検証
- env -u VITE_FIREBASE_API_KEY VITE_FIREBASE_AUTH_DOMAIN VITE_FIREBASE_PROJECT_ID
  VITE_FIREBASE_APP_ID npm run test → 27 files / 445 tests PASS (CI 環境模倣)
- npm run lint PASS

## 教訓
component から pure helper を export して test から import するパターンは、
component が transitive に env-dependent module を引っ張る場合に CI で壊れる。
pure helper は最初から store/firebase 依存ゼロの独立ファイルに置くべき。
